refactor: library 폴더 삭제
