+++ src/js/constants/API.js
@@ -1,3 +1,3 @@
 const YOUTUBE_BASE_URL = 'https://www.googleapis.com/youtube/v3';
 
-export default YOUTUBE_BASE_URL;
+export { YOUTUBE_BASE_URL };
