+++ src/js/init.js
@@ -1,11 +1,11 @@
-import $ from './library/utils/DOM.js';
+import $ from './utils/DOM.js';
 import handleVideoSearch from './handlers/videoSearch.js';
 import handleVideoSave from './handlers/videoSave.js';
 import { openModal, closeModal } from './viewControllers/app.js';
 import handleLatestKeywordSearch from './handlers/latestKeywordSearch.js';
-import videoInfos from './library/states/videoInfos.js';
-import latestKeywords from './library/states/latestKeywords.js';
-import intersectionObserver from './library/states/intersectionObserver.js';
+import videoInfos from './states/videoInfos.js';
+import latestKeywords from './states/latestKeywords.js';
+import intersectionObserver from './states/intersectionObserver.js';
 
 function initState() {
   videoInfos.init();
