+++ src/js/templates/videoSkeleton.js
